+++ src/vivarium/component.py
@@ -16,7 +16,9 @@ from collections.abc import Callable, Sequence
 from datetime import datetime, timedelta
 from importlib import import_module
 from inspect import signature
-from typing import TYPE_CHECKING, Any, cast
+from typing import TYPE_CHECKING, Any
+from typing import SupportsFloat as Numeric
+from typing import cast
 
 import pandas as pd
 from layered_config_tree import ConfigurationError, LayeredConfigTree
@@ -602,7 +604,7 @@ class Component(ABC):
         data = self.get_data(builder, data_source)
         # TODO update this to use vivarium.types.LookupTableData once we drop
         #  support for Python 3.9
-        if not isinstance(data, (float, int, timedelta, datetime, pd.DataFrame, list, tuple)):
+        if not isinstance(data, (Numeric, timedelta, datetime, pd.DataFrame, list, tuple)):
             raise ConfigurationError(f"Data '{data}' must be a LookupTableData instance.")
 
         if isinstance(data, list):