+++ CHANGELOG.rst
@@ -1,3 +1,15 @@
+**3.2.10 - 12/17/24**
+
+  - Type-hinting: Fix mypy errors in vivarium/framework/components/manager.py
+
+**3.2.9 - 12/17/24**
+
+  - Bugfix: Replace certain cases of int or float types with Numeric
+
+**3.2.8 - 12/16/24**
+
+  - Feature: Update how additional seed is defined in RandomnessManager
+
 **3.2.7 - 12/12/24**
 
   - Type-hinting: Fix mypy errors in vivarium/framework/state_machine.py